+++ r_tracer/src/datatypes/vector3.rs
@@ -1,17 +1,17 @@
 use crate::datatypes::color::Color;
 use crate::datatypes::vector2::Vector2;
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 use rand::{Rng, SeedableRng};
 use rand::rngs::StdRng;
 
 #[derive(Copy, Clone, PartialEq)]
 pub struct Vector3 {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64
+    pub x: f32,
+    pub y: f32,
+    pub z: f32
 }
 impl Vector3 {
-    pub fn new(x: f64, y: f64, z: f64) -> Vector3 {
+    pub fn new(x: f32, y: f32, z: f32) -> Vector3 {
         Vector3 { x: x, y: y, z: z}
     }
     pub fn to_color(self) -> Color {
@@ -34,53 +34,53 @@ impl Vector3 {
         }
     }
 
-    pub fn self_dot(self) -> f64 {
+    pub fn self_dot(self) -> f32 {
         self * self
     }
-    pub fn component_add(self) -> f64 {
+    pub fn component_add(self) -> f32 {
         self.x + self.y + self.z
     }
     pub fn square(self) -> Vector3 {
         Vector3::new(self.x * self.x, self.y * self.y, self.z * self.z)
     }
-    pub fn distance(self, other: Vector3) -> f64 {
+    pub fn distance(self, other: Vector3) -> f32 {
         (self - other).square().component_add().sqrt()
     }
     pub fn min(self, other: Vector3) -> Vector3 {
         Vector3::new(
-            f64::min(self.x, other.x),
-             f64::min(self.y, other.y), 
-             f64::min(self.z, other.z)
+            f32::min(self.x, other.x),
+             f32::min(self.y, other.y), 
+             f32::min(self.z, other.z)
         )
     }
     pub fn max(self, other: Vector3) -> Vector3 {
         Vector3::new(
-            f64::max(self.x, other.x),
-             f64::max(self.y, other.y), 
-             f64::max(self.z, other.z)
+            f32::max(self.x, other.x),
+             f32::max(self.y, other.y), 
+             f32::max(self.z, other.z)
         )
     }
-    pub fn component_ave(self) -> f64 {
+    pub fn component_ave(self) -> f32 {
         (self.x + self.y + self.z) / 3.0
     }
-    pub fn rot_x(self, x_degrees: f64) -> Vector3 {
-        let theta_x: f64 = x_degrees*2.0*PI/360.0;
+    pub fn rot_x(self, x_degrees: f32) -> Vector3 {
+        let theta_x: f32 = x_degrees*2.0*PI/360.0;
         Vector3::new(
             self.x,
             self.y*theta_x.cos() - self.z*theta_x.sin(),
             self.y*theta_x.sin() + self.z*theta_x.cos()
         )
     }
-    pub fn rot_y(self, y_degrees: f64) -> Vector3 {
-        let theta_y: f64 = y_degrees*2.0*PI/360.0;
+    pub fn rot_y(self, y_degrees: f32) -> Vector3 {
+        let theta_y: f32 = y_degrees*2.0*PI/360.0;
         Vector3::new(
             self.x*theta_y.cos() + self.z*theta_y.sin(),
             self.y,
             self.z*theta_y.cos() - self.x*theta_y.sin()
         )
     }
-    pub fn rot_z(self, z_degrees: f64) -> Vector3 {
-        let theta_z: f64 = z_degrees*2.0*PI/360.0;
+    pub fn rot_z(self, z_degrees: f32) -> Vector3 {
+        let theta_z: f32 = z_degrees*2.0*PI/360.0;
         Vector3::new(
             self.x*theta_z.cos() - self.y*theta_z.sin(),
             self.x*theta_z.sin() + self.y*theta_z.cos(),
@@ -100,18 +100,18 @@ impl Vector3 {
     pub fn random_hemisphere_normal(normal: Vector3) -> Vector3 {
         let mut rng = rand::thread_rng();
         let mut random_vector = Vector3::new(
-            rng.gen::<f64>() * 2.0 - 1.0,
-            rng.gen::<f64>() * 2.0 - 1.0,
-            rng.gen::<f64>() * 2.0 - 1.0,
+            rng.gen::<f32>() * 2.0 - 1.0,
+            rng.gen::<f32>() * 2.0 - 1.0,
+            rng.gen::<f32>() * 2.0 - 1.0,
         ).normalize();
         if random_vector * normal < 0.0 { random_vector = -1.0 * random_vector }
     
         random_vector
     }
     pub fn random_perturb(scale: Vector2) -> Vector3 {
-        let rand_val1: f64 = rand::thread_rng().gen_range(0.0..1.0);
-        let rand_val2: f64 = rand::thread_rng().gen_range(0.0..1.0);
-        let angle: f64 = rand_val1*2.0*PI;
+        let rand_val1: f32 = rand::thread_rng().gen_range(0.0..1.0);
+        let rand_val2: f32 = rand::thread_rng().gen_range(0.0..1.0);
+        let angle: f32 = rand_val1*2.0*PI;
         let circle_pt: Vector2 = Vector2::new(angle.cos(), angle.sin());
         rand_val2.sqrt()*Vector3::new(0.0, circle_pt.y/scale.x, circle_pt.x/scale.y)
     }
@@ -127,7 +127,7 @@ impl Vector3 {
     }
 
     pub fn normalize(self) -> Vector3 {
-        let magnitude: f64 = self.magnitude();
+        let magnitude: f32 = self.magnitude();
 
         if magnitude != 0.0 {
             Vector3::new(
@@ -144,15 +144,15 @@ impl Vector3 {
         Vector3::new(self.x*other.x, self.y*other.y, self.z*other.z)
     }
 
-    pub fn lerp(v1: Vector3, v2: Vector3, t: f64) -> Vector3 {
+    pub fn lerp(v1: Vector3, v2: Vector3, t: f32) -> Vector3 {
         (1.0 - t)*v1 + t*v2
     }
 
-    pub fn magnitude_squared(self) -> f64 {
+    pub fn magnitude_squared(self) -> f32 {
         self.square().component_add()
     }
 
-    pub fn magnitude(self) -> f64 {
+    pub fn magnitude(self) -> f32 {
         self.magnitude_squared().sqrt()
     }
 
@@ -198,14 +198,14 @@ impl std::ops::Sub<Vector3> for Vector3 {
 }
 
 impl std::ops::Mul<Vector3> for Vector3 {
-    type Output = f64;
+    type Output = f32;
 
-    fn mul(self, other: Vector3) -> f64 {
+    fn mul(self, other: Vector3) -> f32 {
         self.x * other.x + self.y * other.y + self.z * other.z
     }
 }
 
-impl std::ops::Mul<Vector3> for f64 {
+impl std::ops::Mul<Vector3> for f32 {
     type Output = Vector3;
 
     fn mul(self, other: Vector3) -> Vector3 {
@@ -213,24 +213,24 @@ impl std::ops::Mul<Vector3> for f64 {
     }
 }
 
-impl std::ops::Mul<f64> for Vector3 {
+impl std::ops::Mul<f32> for Vector3 {
     type Output = Vector3;
 
-    fn mul(self, other: f64) -> Vector3 {
+    fn mul(self, other: f32) -> Vector3 {
         Vector3::new(self.x * other, self.y * other, self.z * other)
     }
 }
 
-impl std::ops::Div<f64> for Vector3 {
+impl std::ops::Div<f32> for Vector3 {
     type Output = Vector3;
 
-    fn div(self, other: f64) -> Vector3 {
+    fn div(self, other: f32) -> Vector3 {
         Vector3::new(self.x / other, self.y / other, self.z / other)
     }
 }
 
 impl std::ops::Index<usize> for Vector3 {
-    type Output = f64;
+    type Output = f32;
 
     fn index(&self, index: usize) -> &Self::Output {
         match index {
